Safari-611.1.10/LayoutTests/accessibility/mac/label-element-with-link-string-value-expected.txt

 first choice
This tests that if a label element contains a link it will be exposed as a group with no AXValue.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS label.role is 'AXRole: AXGroup'
PASS label.stringValue is 'AXValue: '
PASS successfullyParsed is true

TEST COMPLETE
